Migrate blob-storage-databus to NServiceBus 10
- Updated all .csproj files to target .NET 10 and NServiceBus 10.0.0-alpha.1
- Updated to NServiceBus.ClaimCheck 2.0.0-alpha.1 package
- Updated NServiceBus.DataBus.AzureBlobStorage to 7.0.0-alpha.1
- Updated NServiceBus.Extensions.Hosting to 4.0.0-alpha.1
- Added prerelease.txt marker file
- Removed InputLoopService pattern and moved logic to Program.cs
- Sample already used ClaimCheckProperty<T> so no message migration needed
- Sample already used AzureClaimCheck configuration

Fixes: #7287
